fix: EXP-032 — CallNativeEntry return value propagation (int→long + CpuContext update)
ROOT CAUSE: ExecuteGuestThreadEntry never wrote nativeReturn to CpuContext.Rax.
CallNativeEntry returned int (32-bit), truncating 64-bit RAX.
TryCallGuestFunction read context[Rax] (still 0) instead of nativeReturn.

Fix:
1. CallNativeEntry: int → long (preserve full 64-bit RAX)
2. ExecuteGuestThreadEntry: context[CpuRegister.Rax] = (ulong)nativeReturn
3. NativeWorker.cs: cast CallNativeEntry to int (worker path)
4. Main entry: num6 changed from int to long

Result:
- 232/232 resolver calls now return correct func_impl (was 0/232)
- il2cpp_init → 0x804ED85D0 ✅
- 64-bit return value preserved ✅
- Game progresses past IL2CPP init to next blocker

+++ src/SharpEmu.Core/Cpu/Native/DirectExecutionBackend.NativeWorker.cs
@@ -62,7 +62,7 @@ public sealed partial class DirectExecutionBackend
 		if (worker is null)
 		{
 			TlsSetValue(_hostRspSlotTlsIndex, (nint)hostRspSlot);
-			return CallNativeEntry(entryStub);
+			return (int)CallNativeEntry(entryStub);
 		}
 		try
 		{
